remove HOT details from era5.py
